Move Historical components out of /pages
