+++ vision-aid-prototype-v1/comps/HistoricalCLVForm.js
@@ -8,7 +8,7 @@ import { useSession, signIn, signOut, getSession } from "next-auth/react";
 import { useState, useEffect } from "react";
 import moment from "moment";
 
-export default function HistoricalCLVForm(props={evaluationData:{service:{}, editable: false}}) {
+export default function HistoricalCLVForm(props) {
   // let data = props.evaluationData.service;
   const [data, setData] = useState({});
   useEffect(() => {
