+++ vision-aid-prototype-v1/comps/HistoricalCounselingForm.js
@@ -8,8 +8,8 @@ import { useSession, signIn, signOut, getSession } from "next-auth/react";
 import { useState, useEffect } from "react";
 import moment from "moment";
 
-export default function HistoricalTrainingForm(props={evaluationData:{service:{}, editable: false}}) {
-  // const data = props.evaluationData.service;
+export default function HistoricalCounselingForm(props) {
+  // const data = props.evaluationData;
   const [data, setData] = useState({});
   useEffect(() => {
     setData(props.evaluationData.service);
@@ -26,6 +26,7 @@ export default function HistoricalTrainingForm(props={evaluationData:{service:{}
         ...data,
         [e.target.name]: new Date(Date.parse(e.target.value)),
       });
+      // add reload to include this new date in drop-down
     } else if (e.target.type === "number") {
       setData({ ...data, [e.target.name]: parseInt(e.target.value) });
     } else {
@@ -33,8 +34,8 @@ export default function HistoricalTrainingForm(props={evaluationData:{service:{}
     }
   };
 
-  const saveTrainingData = async () => {
-    const res = await fetch("/api/training", {
+  const saveCounselingData = async () => {
+    const res = await fetch("/api/counsellingEducation", {
       method: "PATCH",
       headers: {
         "Content-Type": "application/json",
@@ -65,7 +66,7 @@ export default function HistoricalTrainingForm(props={evaluationData:{service:{}
           <tr>
             <th scope="row">Date</th>
             <td>
-              {!editMode && data.date !== null && moment(data.date).format("YYYY-MM-DD")}
+            {!editMode && data.date !== null && moment(data.date).format("YYYY-MM-DD")}
               {!editMode && data.date !== null && ""}
               {editMode && (
                 <input
@@ -105,34 +106,10 @@ export default function HistoricalTrainingForm(props={evaluationData:{service:{}
               )}
             </td>
           </tr>
-          <tr>
-            <th scope="row">Sub Type</th>
-            <td>
-              {!editMode && data.subType}
-              {editMode && (
-                <input
-                  type="text"
-                  name="subType"
-                  value={data.subType}
-                  onChange={(e) => handleChange(e)}
-                />
-              )}
-            </td>
-          </tr>
           {/* <tr>
-            <th scope="row">Other Sub Type</th>
-            <td>
-              {!editMode && data.subTypeOther}
-              {editMode && (
-                <input
-                  type="text"
-                  name="subTypeOther"
-                  value={data.subTypeOther}
-                  onChange={(e) => handleChange(e)}
-                />
-              )}
-            </td>
-          </tr> */}
+            <th scope="row">Other Type</th>
+            <td>{data.typeOther}</td>
+            </tr> */}
           <tr>
             <th scope="row">Extra Information</th>
             <td>
@@ -152,7 +129,7 @@ export default function HistoricalTrainingForm(props={evaluationData:{service:{}
       {props.evaluationData.editable && !editMode && (
         <button onClick={handleClick}>Edit</button>
       )}
-      {editMode && <button onClick={saveTrainingData}>Save</button>}
+      {editMode && <button onClick={saveCounselingData}>Save</button>}
     </div>
   );
 }
