+++ vision-aid-prototype-v1/comps/HistoricalVisionEnhancementForm.js
@@ -8,7 +8,7 @@ import { useSession, signIn, signOut, getSession } from "next-auth/react";
 import { useState, useEffect } from "react";
 import moment from "moment";
 
-export default function HistoricalVisionEnhancementForm(props={evaluationData:{service:{}, editable: false}}) {
+export default function HistoricalVisionEnhancementForm(props) {
   //   let data = props.evaluationData.service;
   const [data, setData] = useState({});
   useEffect(() => {
